# docs/swagger.yaml
basePath: /api
definitions:
  handler.getBlogResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Blog'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.Blog:
    properties:
      creation_date:
        type: string
      text:
        type: string
      title:
        type: string
    required:
    - text
    - title
    type: object
  models.CreateBlogParams:
    properties:
      text:
        type: string
      title:
        type: string
    required:
    - text
    - title
    type: object
  models.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: sign in
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
      summary: sign in
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: sign up
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: sign up
      tags:
      - auth
  /blog:
    get:
      description: gets posts
      parameters:
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getBlogResponse'
      summary: Get blog post
      tags:
      - blog
    post:
      consumes:
      - application/json
      description: creates post
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateBlogParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: Create blog post
      tags:
      - blog
swagger: "2.0"